create namespaces with sidecar injection
