+++ data/tuberapps/namespace.go
@@ -1,0 +1,8 @@
+// Package data is generated
+package data
+
+// Namespace is generated. Returns the default namespace for a new tuber app
+var Namespace = TuberYaml{
+	Filename: "namespace.yaml",
+	Contents: []byte{ 97,112,105,86,101,114,115,105,111,110,58,32,118,49,10,107,105,110,100,58,32,78,97,109,101,115,112,97,99,101,10,109,101,116,97,100,97,116,97,58,10,32,32,110,97,109,101,58,32,123,123,46,110,97,109,101,115,112,97,99,101,125,125,10,32,32,108,97,98,101,108,115,58,10,32,32,32,32,105,115,116,105,111,45,105,110,106,101,99,116,105,111,110,58,32,101,110,97,98,108,101,100, },
+}
