+++ data/tuberapps/role.go
@@ -1,0 +1,8 @@
+// Package data is generated
+package data
+
+// Role is generated. Returns the default role for a new tuber app
+var Role = TuberYaml{
+	Filename: "role.yaml",
+	Contents: []byte{ 107,105,110,100,58,32,82,111,108,101,10,97,112,105,86,101,114,115,105,111,110,58,32,114,98,97,99,46,97,117,116,104,111,114,105,122,97,116,105,111,110,46,107,56,115,46,105,111,47,118,49,98,101,116,97,49,10,109,101,116,97,100,97,116,97,58,10,32,32,110,97,109,101,58,32,116,117,98,101,114,45,97,100,109,105,110,10,32,32,110,97,109,101,115,112,97,99,101,58,32,123,123,46,110,97,109,101,115,112,97,99,101,125,125,10,114,117,108,101,115,58,10,32,32,45,32,97,112,105,71,114,111,117,112,115,58,10,32,32,32,32,45,32,39,42,39,10,32,32,32,32,114,101,115,111,117,114,99,101,115,58,10,32,32,32,32,32,32,45,32,39,42,39,10,32,32,32,32,118,101,114,98,115,58,10,32,32,32,32,32,32,45,32,39,42,39, },
+}
