+++ data/tuberapps/yamls/role.yaml
@@ -1,0 +1,12 @@
+kind: Role
+apiVersion: rbac.authorization.k8s.io/v1beta1
+metadata:
+  name: tuber-admin
+  namespace: {{.namespace}}
+rules:
+  - apiGroups:
+    - '*'
+    resources:
+      - '*'
+    verbs:
+      - '*'
