+++ data/tuberapps/yamls/rolebinding.yaml
@@ -1,0 +1,13 @@
+kind: RoleBinding
+apiVersion: rbac.authorization.k8s.io/v1beta1
+metadata:
+  name: tuber-admin
+  namespace: {{.namespace}}
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: Role
+  name: tuber-admin
+subjects:
+  - kind: ServiceAccount
+    name: default
+    namespace: tuber
