+++ pkg/core/tuber_layers.go
@@ -1,9 +1,7 @@
 package core
 
 import (
-	"bytes"
 	"strings"
-	"text/template"
 )
 
 // ReleaseTubers combines and interpolates with tuber's conventions, and applies them
@@ -23,21 +21,4 @@ func tuberData(digest string, clusterData *ClusterData) (data map[string]string)
 		"clusterDefaultGateway": clusterData.DefaultGateway,
 		"clusterDefaultHost":    clusterData.DefaultHost,
 	}
-}
-
-func interpolateTuber(templateString string, data map[string]string) (interpolated []byte, err error) {
-	tpl, err := template.New("").Parse(templateString)
-
-	if err != nil {
-		return
-	}
-	var buf bytes.Buffer
-	err = tpl.Execute(&buf, data)
-
-	if err != nil {
-		return
-	}
-
-	interpolated = buf.Bytes()
-	return
 }
